Move upload logic to useUploader hook. Update uploader usage in Account components.
